add: insert number

+++ 0x01-python-if_else_loops_functions/lists.h
@@ -1,5 +1,7 @@
 #ifndef LISTS_H
 #define LISTS_H
+#include <stdlib.h>
+#include <unistd.h>
 
 /**
  * struct listint_s - singly linked list
@@ -14,7 +16,7 @@ typedef struct listint_s
 	    struct listint_s *next;
 } listint_t;
 
-unsigned int print_listint(const listint_t *h);
+size_t print_listint(const listint_t *h);
 listint_t *add_nodeint_end(listint_t **head, const int n);
 void free_listint(listint_t *head);
